made ID checker into a singular component
