+++ components/RequestForm.tsx
@@ -2,17 +2,10 @@ import React, { useState } from "react";
 import { regions, roles } from "../util/datalist";
 import { ChampionSelect } from "./Styles";
 import axios from "axios";
-import {
-  ActionIcon,
-  Button,
-  Container,
-  Select,
-  Stack,
-  TextInput,
-  Textarea,
-} from "@mantine/core";
+import { Button, Container, Select, Stack, Textarea } from "@mantine/core";
 import { useForm, isNotEmpty, hasLength } from "@mantine/form";
 import { timeZones } from "../util/helper";
+import { RiotIdCheckerField } from "./FormComponents";
 
 const RequestForm = ({ setSent }) => {
   const { isValid, values, getInputProps, setValues, setErrors, errors } =
@@ -64,29 +57,13 @@ const RequestForm = ({ setSent }) => {
         <Stack>
           <Select label="Region" data={regions} {...getInputProps("region")} />
 
-          <TextInput
-            label="Riot ID"
-            placeholder={
-              values.region
-                ? "Your full Riot ID with the # code"
-                : "Enter your region first"
-            }
-            description={
-              values.summonerName &&
-              !values.rank &&
-              "Check if your account is valid with button on the right"
-            }
-            rightSection={
-              <RiotIdChecker
-                values={values}
-                error={errors.summonerName}
-                setValues={setValues}
-                setErrors={setErrors}
-                isApply={false}
-              />
-            }
-            disabled={!values.region}
-            {...getInputProps("summonerName")}
+          <RiotIdCheckerField
+            isApply={false}
+            values={values}
+            errors={errors}
+            setValues={setValues}
+            setErrors={setErrors}
+            getInputProps={getInputProps}
           />
 
           <Select label="Role" data={roles} {...getInputProps("role")} />
@@ -110,52 +87,3 @@ const RequestForm = ({ setSent }) => {
   );
 };
 export default RequestForm;
-
-export const RiotIdChecker = ({
-  values: { summonerName, region, rank },
-  setValues,
-  setErrors,
-  error,
-  isApply,
-}) => {
-  const [loading, setLoading] = useState(false);
-
-  const checkAccount = () => {
-    setLoading(true);
-    setErrors({ summonerName: false });
-    axios
-      .put("/api/request", { riotId: summonerName, region })
-      .then(({ data }) => {
-        const rank = data.tier;
-        const games = data.wins + data.losses;
-        setLoading(false);
-        if ((isApply && games) || games >= 25) setValues({ rank });
-        else {
-          setValues({ rank: null });
-          setErrors({
-            summonerName: isApply
-              ? "We are unable to verify your rank. Please make sure you have the correct account or you are ranked"
-              : "You do not meet the required number of games. Please apply after meeting the requirements",
-          });
-        }
-      })
-      .catch(() => {
-        setValues({ rank: null });
-        setErrors({
-          summonerName:
-            "An error has occurred. Please check if you entered the correct ID",
-        });
-        setLoading(false);
-      });
-  };
-  return (
-    <ActionIcon
-      loading={loading}
-      onClick={checkAccount}
-      variant={"filled"}
-      color={rank ? "green" : error ? "red" : "blue"}
-    >
-      {loading ? null : error ? "!" : "O"}
-    </ActionIcon>
-  );
-};
